add double processing

+++ src/task6/Matrix.java
@@ -1,19 +1,19 @@
 package task6;
 
 public class Matrix {
-    private int[][] matrix;
+    private double[][] matrix;
     private int size;
-    public Matrix(int[] array){
+    public Matrix(double[] array){
         size = array.length;
-        matrix = new int[size][size];
+        matrix = new double[size][size];
         fillMatrix(array);
     }
 
-    public int[][] get() {
+    public double[][] get() {
         return matrix;
     }
 
-    public void fillMatrix(int[] array) {
+    public void fillMatrix(double[] array) {
         int startInd;
         for(int i=0; i<size; i++) {
             startInd = i;
@@ -27,7 +27,7 @@ public class Matrix {
     public void print() {
         for(int i=0; i<size; i++) {
             for(int j=0; j<size; j++) {
-                System.out.format("%d ", matrix[i][j]);
+                System.out.format("%.2f ", matrix[i][j]);
             }
             System.out.println();
         }
